created reducer and extracted it plus helpers to a separate file
